Delete alert progress

+++ ExpressServer/routes/deleteAlert.js
@@ -5,7 +5,8 @@ var db = require('./../../database/dbQuery');
 router.post('/', function(req, res, next){
 	const alert = req.body 
 	// Tell mongo what to delete
-	db.deleteAlert( alert.name );
+	db.deleteAlert( alert.username, alert.name ).then(function(confirmation){
+	console.log(confirmation);});
 	
 });//end of POST
